requirements.txt
django>=5.0,<7.0
djangorestframework>=3.14
django-filter>=23.0
gunicorn>=21.0
whitenoise>=6.5
dj-database-url>=2.1
psycopg2-binary>=2.9
